requirement.txt
sentence-transformers 
faiss-cpu 
PyPDF2 
python-docx
fastapi
uvicorn
requests
python-dotenv
numpy